+++ sdk/devcenter/azure-resourcemanager-devcenter/src/main/java/com/azure/resourcemanager/devcenter/models/Catalogs.java
@@ -7,12 +7,15 @@ package com.azure.resourcemanager.devcenter.models;
 import com.azure.core.http.rest.PagedIterable;
 import com.azure.core.http.rest.Response;
 import com.azure.core.util.Context;
+import com.azure.resourcemanager.devcenter.fluent.models.CatalogInner;
 
-/** Resource collection API of Catalogs. */
+/**
+ * Resource collection API of Catalogs.
+ */
 public interface Catalogs {
     /**
      * Lists catalogs for a devcenter.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
@@ -24,7 +27,7 @@ public interface Catalogs {
 
     /**
      * Lists catalogs for a devcenter.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @param top The maximum number of resources to return from the operation. Example: '$top=10'.
@@ -34,12 +37,12 @@ public interface Catalogs {
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return results of the catalog list operation as paginated response with {@link PagedIterable}.
      */
-    PagedIterable<Catalog> listByDevCenter(
-        String resourceGroupName, String devCenterName, Integer top, Context context);
+    PagedIterable<Catalog> listByDevCenter(String resourceGroupName, String devCenterName, Integer top,
+        Context context);
 
     /**
      * Gets a catalog.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @param catalogName The name of the Catalog.
@@ -49,12 +52,12 @@ public interface Catalogs {
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return a catalog along with {@link Response}.
      */
-    Response<Catalog> getWithResponse(
-        String resourceGroupName, String devCenterName, String catalogName, Context context);
+    Response<Catalog> getWithResponse(String resourceGroupName, String devCenterName, String catalogName,
+        Context context);
 
     /**
      * Gets a catalog.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @param catalogName The name of the Catalog.
@@ -66,8 +69,68 @@ public interface Catalogs {
     Catalog get(String resourceGroupName, String devCenterName, String catalogName);
 
     /**
+     * Creates or updates a catalog.
+     * 
+     * @param resourceGroupName The name of the resource group. The name is case insensitive.
+     * @param devCenterName The name of the devcenter.
+     * @param catalogName The name of the Catalog.
+     * @param body Represents a catalog.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return represents a catalog.
+     */
+    Catalog createOrUpdate(String resourceGroupName, String devCenterName, String catalogName, CatalogInner body);
+
+    /**
+     * Creates or updates a catalog.
+     * 
+     * @param resourceGroupName The name of the resource group. The name is case insensitive.
+     * @param devCenterName The name of the devcenter.
+     * @param catalogName The name of the Catalog.
+     * @param body Represents a catalog.
+     * @param context The context to associate with this operation.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return represents a catalog.
+     */
+    Catalog createOrUpdate(String resourceGroupName, String devCenterName, String catalogName, CatalogInner body,
+        Context context);
+
+    /**
+     * Partially updates a catalog.
+     * 
+     * @param resourceGroupName The name of the resource group. The name is case insensitive.
+     * @param devCenterName The name of the devcenter.
+     * @param catalogName The name of the Catalog.
+     * @param body Updatable catalog properties.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return represents a catalog.
+     */
+    Catalog update(String resourceGroupName, String devCenterName, String catalogName, CatalogUpdate body);
+
+    /**
+     * Partially updates a catalog.
+     * 
+     * @param resourceGroupName The name of the resource group. The name is case insensitive.
+     * @param devCenterName The name of the devcenter.
+     * @param catalogName The name of the Catalog.
+     * @param body Updatable catalog properties.
+     * @param context The context to associate with this operation.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return represents a catalog.
+     */
+    Catalog update(String resourceGroupName, String devCenterName, String catalogName, CatalogUpdate body,
+        Context context);
+
+    /**
      * Deletes a catalog resource.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @param catalogName The name of the Catalog.
@@ -79,7 +142,7 @@ public interface Catalogs {
 
     /**
      * Deletes a catalog resource.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @param catalogName The name of the Catalog.
@@ -92,7 +155,7 @@ public interface Catalogs {
 
     /**
      * Gets catalog synchronization error details.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @param catalogName The name of the Catalog.
@@ -102,12 +165,12 @@ public interface Catalogs {
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return catalog synchronization error details along with {@link Response}.
      */
-    Response<SyncErrorDetails> getSyncErrorDetailsWithResponse(
-        String resourceGroupName, String devCenterName, String catalogName, Context context);
+    Response<SyncErrorDetails> getSyncErrorDetailsWithResponse(String resourceGroupName, String devCenterName,
+        String catalogName, Context context);
 
     /**
      * Gets catalog synchronization error details.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @param catalogName The name of the Catalog.
@@ -120,7 +183,7 @@ public interface Catalogs {
 
     /**
      * Syncs templates for a template source.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @param catalogName The name of the Catalog.
@@ -132,7 +195,7 @@ public interface Catalogs {
 
     /**
      * Syncs templates for a template source.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @param catalogName The name of the Catalog.
@@ -145,7 +208,7 @@ public interface Catalogs {
 
     /**
      * Connects a catalog to enable syncing.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @param catalogName The name of the Catalog.
@@ -157,7 +220,7 @@ public interface Catalogs {
 
     /**
      * Connects a catalog to enable syncing.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param devCenterName The name of the devcenter.
      * @param catalogName The name of the Catalog.
@@ -167,56 +230,4 @@ public interface Catalogs {
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      */
     void connect(String resourceGroupName, String devCenterName, String catalogName, Context context);
-
-    /**
-     * Gets a catalog.
-     *
-     * @param id the resource ID.
-     * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
-     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return a catalog along with {@link Response}.
-     */
-    Catalog getById(String id);
-
-    /**
-     * Gets a catalog.
-     *
-     * @param id the resource ID.
-     * @param context The context to associate with this operation.
-     * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
-     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return a catalog along with {@link Response}.
-     */
-    Response<Catalog> getByIdWithResponse(String id, Context context);
-
-    /**
-     * Deletes a catalog resource.
-     *
-     * @param id the resource ID.
-     * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
-     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     */
-    void deleteById(String id);
-
-    /**
-     * Deletes a catalog resource.
-     *
-     * @param id the resource ID.
-     * @param context The context to associate with this operation.
-     * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
-     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     */
-    void deleteByIdWithResponse(String id, Context context);
-
-    /**
-     * Begins definition for a new Catalog resource.
-     *
-     * @param name resource name.
-     * @return the first stage of the new Catalog definition.
-     */
-    Catalog.DefinitionStages.Blank define(String name);
 }