+++ sdk/hybridcompute/azure-resourcemanager-hybridcompute/src/main/java/com/azure/resourcemanager/hybridcompute/models/PublicNetworkAccessType.java
@@ -13,15 +13,19 @@ import java.util.Collection;
  * disabled (access to Azure Arc services only via private link).
  */
 public final class PublicNetworkAccessType extends ExpandableStringEnum<PublicNetworkAccessType> {
-    /** Static value Enabled for PublicNetworkAccessType. */
+    /**
+     * Static value Enabled for PublicNetworkAccessType.
+     */
     public static final PublicNetworkAccessType ENABLED = fromString("Enabled");
 
-    /** Static value Disabled for PublicNetworkAccessType. */
+    /**
+     * Static value Disabled for PublicNetworkAccessType.
+     */
     public static final PublicNetworkAccessType DISABLED = fromString("Disabled");
 
     /**
      * Creates a new instance of PublicNetworkAccessType value.
-     *
+     * 
      * @deprecated Use the {@link #fromString(String)} factory method.
      */
     @Deprecated
@@ -30,7 +34,7 @@ public final class PublicNetworkAccessType extends ExpandableStringEnum<PublicNe
 
     /**
      * Creates or finds a PublicNetworkAccessType from its string representation.
-     *
+     * 
      * @param name a name to look for.
      * @return the corresponding PublicNetworkAccessType.
      */
@@ -41,7 +45,7 @@ public final class PublicNetworkAccessType extends ExpandableStringEnum<PublicNe
 
     /**
      * Gets known PublicNetworkAccessType values.
-     *
+     * 
      * @return known PublicNetworkAccessType values.
      */
     public static Collection<PublicNetworkAccessType> values() {