+++ sdk/support/azure-resourcemanager-support/src/main/java/com/azure/resourcemanager/support/implementation/FileWorkspacesNoSubscriptionsClientImpl.java
@@ -52,8 +52,8 @@ public final class FileWorkspacesNoSubscriptionsClientImpl implements FileWorksp
     }
 
     /**
-     * The interface defining all the services for MicrosoftSupportFileWorkspacesNoSubscriptions to be used by the
-     * proxy service to perform REST calls.
+     * The interface defining all the services for MicrosoftSupportFileWorkspacesNoSubscriptions to be used by the proxy
+     * service to perform REST calls.
      */
     @Host("{$host}")
     @ServiceInterface(name = "MicrosoftSupportFile")